+++ packages/create-app/index.js
@@ -11,6 +11,7 @@ const {
   cyan,
   magenta,
   lightRed,
+  red,
   stripColors
 } = require('kolorist')
 
@@ -25,7 +26,9 @@ const TEMPLATES = [
   magenta('preact'),
   magenta('preact-ts'),
   lightRed('lit-element'),
-  lightRed('lit-element-ts')
+  lightRed('lit-element-ts'),
+  red('svelte'),
+  red('svelte-ts')
 ]
 
 const renameFiles = {
@@ -36,17 +39,17 @@ async function init() {
   let targetDir = argv._[0]
   if (!targetDir) {
     /**
-     * @type {{ name: string }}
+     * @type {{ projectName: string }}
      */
-    const { name } = await prompt({
+    const { projectName } = await prompt({
       type: 'input',
-      name: 'name',
+      name: 'projectName',
       message: `Project name:`,
       initial: 'vite-project'
     })
-    targetDir = name
+    targetDir = projectName
   }
-
+  const packageName = await getValidPackageName(targetDir)
   const root = path.join(cwd, targetDir)
   console.log(`\nScaffolding project in ${root}...`)
 
@@ -118,7 +121,9 @@ async function init() {
   }
 
   const pkg = require(path.join(templateDir, `package.json`))
-  pkg.name = path.basename(root)
+
+  pkg.name = packageName
+
   write('package.json', JSON.stringify(pkg, null, 2))
 
   const pkgManager = /yarn/.test(process.env.npm_execpath) ? 'yarn' : 'npm'
@@ -138,6 +143,33 @@ function copy(src, dest) {
     copyDir(src, dest)
   } else {
     fs.copyFileSync(src, dest)
+  }
+}
+
+async function getValidPackageName(projectName) {
+  const packageNameRegExp = /^(?:@[a-z0-9-*~][a-z0-9-*._~]*\/)?[a-z0-9-~][a-z0-9-._~]*$/
+  if (packageNameRegExp.test(projectName)) {
+    return projectName
+  } else {
+    const suggestedPackageName = projectName
+      .trim()
+      .toLowerCase()
+      .replace(/\s+/g, '-')
+      .replace(/^[._]/, '')
+      .replace(/[^a-z0-9-~]+/g, '-')
+
+    /**
+     * @type {{ inputPackageName: string }}
+     */
+    const { inputPackageName } = await prompt({
+      type: 'input',
+      name: 'inputPackageName',
+      message: `Package name:`,
+      initial: suggestedPackageName,
+      validate: (input) =>
+        packageNameRegExp.test(input) ? true : 'Invalid package.json name'
+    })
+    return inputPackageName
   }
 }