+++ packages/vite/src/node/server/hmr.ts
@@ -7,7 +7,9 @@ import { ModuleNode } from './moduleGraph'
 import { Update } from 'types/hmrPayload'
 import { CLIENT_DIR } from '../constants'
 import { RollupError } from 'rollup'
+import { prepareError } from './middlewares/error'
 import match from 'minimatch'
+import { Server } from 'http'
 
 export const debugHmr = createDebugger('vite:hmr')
 
@@ -20,6 +22,7 @@ export interface HmrOptions {
   path?: string
   timeout?: number
   overlay?: boolean
+  server?: Server
 }
 
 export interface HmrContext {
@@ -162,10 +165,10 @@ export async function handleFileAddUnlink(
   server: ViteDevServer,
   isUnlink = false
 ) {
+  const modules = [...(server.moduleGraph.getModulesByFile(file) ?? [])]
   if (isUnlink && file in server._globImporters) {
     delete server._globImporters[file]
   } else {
-    const modules = []
     for (const i in server._globImporters) {
       const { module, base, pattern } = server._globImporters[i]
       const relative = path.relative(base, file)
@@ -173,14 +176,14 @@ export async function handleFileAddUnlink(
         modules.push(module)
       }
     }
-    if (modules.length > 0) {
-      updateModules(
-        getShortName(file, server.config.root),
-        modules,
-        Date.now(),
-        server
-      )
-    }
+  }
+  if (modules.length > 0) {
+    updateModules(
+      getShortName(file, server.config.root),
+      modules,
+      Date.now(),
+      server
+    )
   }
 }
 
@@ -234,7 +237,11 @@ function invalidate(mod: ModuleNode, timestamp: number, seen: Set<ModuleNode>) {
   seen.add(mod)
   mod.lastHMRTimestamp = timestamp
   mod.transformResult = null
-  mod.importers.forEach((importer) => invalidate(importer, timestamp, seen))
+  mod.importers.forEach((importer) => {
+    if (!importer.acceptedHmrDeps.has(mod)) {
+      invalidate(importer, timestamp, seen)
+    }
+  })
 }
 
 export function handlePrunedModules(
@@ -410,9 +417,20 @@ async function readModifiedFile(file: string): Promise<string> {
 }
 
 async function restartServer(server: ViteDevServer) {
+  // @ts-ignore
+  global.__vite_start_time = Date.now()
+  let newServer = null
+  try {
+    newServer = await createServer(server.config.inlineConfig)
+  } catch (err) {
+    server.ws.send({
+      type: 'error',
+      err: prepareError(err)
+    })
+    return
+  }
+
   await server.close()
-  ;(global as any).__vite_start_time = Date.now()
-  const newServer = await createServer(server.config.inlineConfig)
   for (const key in newServer) {
     if (key !== 'app') {
       // @ts-ignore